pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling>=1.26"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
authors = [
    { name = "Teemu Viikeri", email = "tepavi@gmail.com" },
]
description = "An interactive command-line tool for GitLab REST API."
keywords = ["gitlab"]
name = "glcli"
readme = "README.md"
version = "0.1.0"

dependencies = [
    "python-gitlab>=5.6.0",
    "tomli>=2.2.1",
    "typer>=0.15.2",
    "yaspin>=3.1.0",
]
requires-python = ">=3.9"

[project.scripts]
glcli = "src.main:app"

[dependency-groups]
dev = [
    "ruff>=0.11.7",
]
